Manipulate the sizes of the leaves with a Weibull distribution

This function returns a function which itself returns a numerical vector of length of the number of leaves on the branch.

spark_weibull(shape = 1.2, scale_factor = 0.5)

Arguments

shape, scale_factor
Parameters passed to stats::dweibull().

Value

dweibull() function with n_leaves as one of the arguments

Examples

spark_weibull()

#> function (n_leaves) 
#> {
#>     stats::dweibull(1:n_leaves, shape, scale = n_leaves * scale_factor)
#> }
#> <bytecode: 0x55dd1af67b60>
#> <environment: 0x55dd1d7d7ba0>